items: rename shufflers
