+++ BlasII.Randomizer/Items/Shuffle/PoolsItemShuffler.cs
@@ -1,22 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace BlasII.Randomizer.Items
+namespace BlasII.Randomizer.Items.Shuffle
 {
-    internal class ItemShufflerReverse : IShuffler
+    internal class PoolsItemShuffler : IShuffler
     {
         public bool Shuffle(int seed, RandomizerSettings settings, Dictionary<string, string> output)
         {
             output.Clear();
-            Initialize(seed);
+            var rng = new Random(seed);
 
             // Create pools of all locations to randomize
-            List<ItemLocation> progressionLocations = new(), junkLocations = new();
+            LocationPool progressionLocations = new(rng), junkLocations = new(rng);
             CreateLocationPool(progressionLocations, junkLocations, settings);
 
             // Create pools of all items to randomize
-            List<Item> progressionItems = new(), junkItems = new();
-            CreateItemPool(progressionItems, junkItems, progressionLocations.Count + junkLocations.Count, settings);
+            ItemPool progressionItems = new(rng), junkItems = new(rng);
+            CreateItemPool(progressionItems, junkItems, progressionLocations.Size + junkLocations.Size, settings);
 
             // Create initial inventory
             var inventory = new Blas2Inventory(settings, Main.Randomizer.Data.DoorDictionary);
@@ -26,15 +26,15 @@ namespace BlasII.Randomizer.Items
             FillProgressionItems(progressionLocations, progressionItems, output, inventory);
 
             // Add junk items to remaining progression items and ensure they are still balanced
-            junkLocations.AddRange(progressionLocations);
-            if (junkLocations.Count != junkItems.Count)
+            junkLocations.Combine(progressionLocations);
+            if (junkLocations.Size != junkItems.Size)
                 return false;
 
             // Place junk items at junk locations and remaining progression locations
             FillJunkItems(junkLocations, junkItems, output);
 
             // Verify that all remaining items were placed
-            return junkItems.Count == 0 && junkLocations.Count == 0;
+            return junkItems.Size == 0 && junkLocations.Size == 0;
         }
 
         #region Setup
@@ -42,7 +42,7 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// Fills the two location pools
         /// </summary>
-        private void CreateLocationPool(List<ItemLocation> progressionLocations, List<ItemLocation> junkLocations, RandomizerSettings settings)
+        private void CreateLocationPool(LocationPool progressionLocations, LocationPool junkLocations, RandomizerSettings settings)
         {
             foreach (var location in Main.Randomizer.Data.ItemLocationList)
             {
@@ -53,16 +53,16 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// Takes a single location data and adds it to the correct list based on its type
         /// </summary>
-        private void AddLocationToPool(List<ItemLocation> progressionLocations, List<ItemLocation> junkLocations, ItemLocation location, RandomizerSettings settings)
+        private void AddLocationToPool(LocationPool progressionLocations, LocationPool junkLocations, ItemLocation location, RandomizerSettings settings)
         {
-            List<ItemLocation> locationPool = location.ShouldBeShuffled(settings) ? progressionLocations : junkLocations;
+            LocationPool locationPool = location.ShouldBeShuffled(settings) ? progressionLocations : junkLocations;
             locationPool.Add(location);
         }
 
         /// <summary>
         /// Fills the two item pools to match the number of locations
         /// </summary>
-        private void CreateItemPool(List<Item> progressionItems, List<Item> junkItems, int numOfLocations, RandomizerSettings settings)
+        private void CreateItemPool(ItemPool progressionItems, ItemPool junkItems, int numOfLocations, RandomizerSettings settings)
         {
             foreach (var item in Main.Randomizer.Data.ItemList)
             {
@@ -76,9 +76,9 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// Takes a single item data and adds it to the correct list based on its type
         /// </summary>
-        private void AddItemToPool(List<Item> progressionItems, List<Item> junkItems, Item item)
+        private void AddItemToPool(ItemPool progressionItems, ItemPool junkItems, Item item)
         {
-            List<Item> itemPool = item.progression ? progressionItems : junkItems;
+            ItemPool itemPool = item.progression ? progressionItems : junkItems;
             for (int i = 0; i < item.count; i++)
             {
                 itemPool.Add(item);
@@ -88,7 +88,7 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// Removes the starting weapon from the item pool
         /// </summary>
-        private void RemoveStartingItemsFromItemPool(List<Item> items, RandomizerSettings settings)
+        private void RemoveStartingItemsFromItemPool(ItemPool items, RandomizerSettings settings)
         {
             // Remove the extra starting weapon
             items.Remove(Main.Randomizer.Data.GetItem(GetStartingWeaponId(settings)));
@@ -97,16 +97,16 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// After creating the item pool, add or remove junk items to make it equal to the number of locations
         /// </summary>
-        private void BalanceItemPool(List<Item> progressionItems, List<Item> junkItems, int numOfLocations)
+        private void BalanceItemPool(ItemPool progressionItems, ItemPool junkItems, int numOfLocations)
         {
             // Remove tear items until pools are equal
-            while (progressionItems.Count + junkItems.Count > numOfLocations)
+            while (progressionItems.Size + junkItems.Size > numOfLocations)
             {
-                junkItems.RemoveAt(junkItems.Count - 1);
+                junkItems.RemoveLast();
             }
 
             // Add tear items until pools are equal
-            while (progressionItems.Count + junkItems.Count < numOfLocations)
+            while (progressionItems.Size + junkItems.Size < numOfLocations)
             {
                 junkItems.Add(Main.Randomizer.Data.GetItem("Tears[800]"));
             }
@@ -115,7 +115,7 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// Adds the starting weapon and all progression items to the initial inventory
         /// </summary>
-        private void CreateInitialInventory(Blas2Inventory inventory, RandomizerSettings settings, List<Item> progressionItems)
+        private void CreateInitialInventory(Blas2Inventory inventory, RandomizerSettings settings, ItemPool progressionItems)
         {
             // Add the starting weapon
             inventory.AddItem(Main.Randomizer.Data.GetItem(GetStartingWeaponId(settings)));
@@ -142,7 +142,7 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// Using a reverse fill algorithm, place the last item at a random reachable location
         /// </summary>
-        private void FillProgressionItems(List<ItemLocation> locations, List<Item> items, Dictionary<string, string> output, Blas2Inventory inventory)
+        private void FillProgressionItems(LocationPool locations, ItemPool items, Dictionary<string, string> output, Blas2Inventory inventory)
         {
             // Verify that all locations are reachable
             foreach (var location in locations)
@@ -151,20 +151,20 @@ namespace BlasII.Randomizer.Items
                     return;
             }
 
-            ShuffleList(items);
+            items.Shuffle();
             MovePriorityItems(items);
-            var reachableLocations = new List<ItemLocation>(locations);
+            var reachableLocations = new LocationPool(locations);
 
-            while (reachableLocations.Count > 0 && items.Count > 0)
+            while (reachableLocations.Size > 0 && items.Size > 0)
             {
-                Item item = RemoveLast(items);
+                Item item = items.RemoveLast();
                 inventory.RemoveItem(item);
 
                 RemoveUnreachableLocations(reachableLocations, inventory);
-                if (reachableLocations.Count == 0)
+                if (reachableLocations.Size == 0)
                     return;
 
-                ItemLocation location = RemoveRandom(reachableLocations);
+                ItemLocation location = reachableLocations.RemoveRandom();
                 locations.Remove(location);
                 output.Add(location.id, item.id);
             }
@@ -173,14 +173,14 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// Without logic, place the last item at a random location
         /// </summary>
-        private void FillJunkItems(List<ItemLocation> locations, List<Item> items, Dictionary<string, string> output)
+        private void FillJunkItems(LocationPool locations, ItemPool items, Dictionary<string, string> output)
         {
-            ShuffleList(items);
+            items.Shuffle();
 
-            while (locations.Count > 0 && items.Count > 0)
+            while (locations.Size > 0 && items.Size > 0)
             {
-                ItemLocation location = RemoveLast(locations);
-                Item item = RemoveLast(items);
+                ItemLocation location = locations.RemoveLast();
+                Item item = items.RemoveLast();
 
                 output.Add(location.id, item.id);
             }
@@ -189,77 +189,20 @@ namespace BlasII.Randomizer.Items
         /// <summary>
         /// After shuffling the list of progression items, move certain items to the end to prevent failing seeds
         /// </summary>        
-        private void MovePriorityItems(List<Item> progressionItems)
+        private void MovePriorityItems(ItemPool progressionItems)
         {
             Item wallClimb = Main.Randomizer.Data.GetItem("AB44");
-            progressionItems.Remove(wallClimb);
-            progressionItems.Insert(0, wallClimb);
+            progressionItems.MoveToBeginning(wallClimb);
         }
 
         /// <summary>
         /// Finds all locations that are no longer reachable and removes them
         /// </summary>
-        private void RemoveUnreachableLocations(List<ItemLocation> locations, Blas2Inventory inventory)
+        private void RemoveUnreachableLocations(LocationPool locations, Blas2Inventory inventory)
         {
-            for (int i = 0; i < locations.Count; i++)
-            {
-                if (inventory.Evaluate(locations[i].logic))
-                    continue;
-
-                locations.RemoveAt(i);
-                i--;
-            }
+            locations.RemoveConditional(loc => !inventory.Evaluate(loc.logic));
         }
 
         #endregion
-
-        // Old base shuffler stuff
-
-        private Random rng;
-
-        protected void Initialize(int seed) => rng = new Random(seed);
-
-        protected int RandomInteger(int max) => rng.Next(max);
-
-        protected T RandomElement<T>(List<T> list) => list[RandomInteger(list.Count)];
-
-        protected void ShuffleList<T>(List<T> list)
-        {
-            int upperIdx = list.Count;
-            while (upperIdx > 1)
-            {
-                upperIdx--;
-                int randIdx = RandomInteger(upperIdx + 1);
-                T value = list[randIdx];
-                list[randIdx] = list[upperIdx];
-                list[upperIdx] = value;
-            }
-        }
-
-        protected T RemoveRandom<T>(List<T> list)
-        {
-            int index = RandomInteger(list.Count);
-            T element = list[index];
-
-            list.RemoveAt(index);
-            return element;
-        }
-
-        protected T RemoveRandomFromOther<T>(List<T> list, List<T> other)
-        {
-            T element = RandomElement(list);
-
-            other.Remove(element);
-            return element;
-        }
-
-        protected T RemoveLast<T>(List<T> list)
-        {
-            int index = list.Count - 1;
-            T element = list[index];
-
-            list.RemoveAt(index);
-            return element;
-        }
     }
 }
